requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.5
APScheduler==3.10.4
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2024.1

# For SAP Ariba automated fetching
selenium==4.15.0
webdriver-manager==4.0.1

gunicorn==21.2.0
